move types definitions to z_balance_streaming.rs mod
