+++ mm2src/coins/z_coin/storage/blockdb/blockdb_sql_storage.rs
@@ -1,8 +1,9 @@
 use crate::z_coin::storage::{scan_cached_block, validate_chain, BlockDbImpl, BlockProcessingMode, CompactBlockRow,
                              ZcoinStorageRes};
 use crate::z_coin::z_coin_errors::ZcoinStorageError;
-use crate::z_coin::{ZBalanceEvent, ZcoinConsensusParams};
+use crate::z_coin::ZcoinConsensusParams;
 
+use crate::z_coin::z_balance_streaming::ZBalanceEvent;
 use common::async_blocking;
 use db_common::sqlite::rusqlite::{params, Connection};
 use db_common::sqlite::{query_single_row, run_optimization_pragmas, rusqlite};
@@ -227,8 +228,8 @@ impl BlockDbImpl {
                 },
                 BlockProcessingMode::Scan(data, z_balance_change_sender) => {
                     let tx_size = scan_cached_block(data, &params, &block, &mut from_height).await?;
-                    // If the current block represents the latest scanned block and there are transactions present,
-                    // we trigger a `Triggered` event to update the balance change.
+                    // If there are transactions present in the current scanned block,
+                    // we send a `Triggered` event to update the balance change.
                     if tx_size > 0 {
                         if let Some(mut sender) = z_balance_change_sender.clone() {
                             sender
